+++ R/data.r
@@ -6,7 +6,8 @@
 #' or object) and the id of the *batch* to which the keyword is assigned.
 #' Keywords can be added with the function `add_keywords`. The function
 #' `start_db` exports the table *batch_keywords* as objects
-#' `keywords_control` and `keywords_object` to the package environment.
+#' `keywords_control` and `keywords_object` to the package environment
+#' `gt.env`.
 #'
 #' Example data for the table *batch_keywords* is available as R object
 #' `example_keywords`.
@@ -64,8 +65,8 @@
 #' gets the value *world* as location. Data is downloaded and automatically
 #' written to the table through the function `download_control`. The
 #' function `start_db` exports the table *data_control* as database
-#' connection `tbl_control` to the package environment. Users can access the
-#' database table through `dplyr::tbl`.
+#' connection `tbl_control` to the package environment `gt.env`. Users
+#' can access the database table through `dplyr::tbl`.
 #' The sample data included in `data_control` was simulated based on actual
 #' Google Trends data.
 #'
@@ -103,8 +104,8 @@
 #' mapped. Global data takes the value *world* as location. Data is
 #' downloaded and automatically written to the table through the function
 #' `download_object`. The function `start_db` exports the table
-#' *data_object* as database connection `tbl_object` to
-#' the package environment. Users can access the database table through
+#' *data_object* as database connection `tbl_object` to the package
+#' environment `gt.env`. Users can access the database table through
 #' `dplyr::tbl`.
 #' The sample data included in `data_object` was simulated based on actual
 #' Google Trends data.
@@ -148,8 +149,8 @@
 #' scores are computed and automatically written to the table with the function
 #' `compute_score`. The function `start_db` exports the table
 #' *data_score* as database connection `tbl_score` to
-#' the package environment. Users can access the database table through
-#' `dplyr::tbl`.
+#' the package environment `gt.env`. Users can access the database
+#' table through `dplyr::tbl`.
 #' The sample data included in `data_score` was simulated based on actual
 #' Google Trends data.
 #'
@@ -199,8 +200,8 @@
 #' distribution of search scores. DOI is computed and automatically written to
 #' the table with the function `compute_doi`. The function `start_db`
 #' exports the table *data_doi* as database connection `tbl_doi` to
-#' the package environment. Users can access the database table through
-#' `dplyr::tbl`.
+#' the package environment `gt.env`. Users can access the database table
+#' through `dplyr::tbl`.
 #' The sample data included in `data_doi` was simulated based on actual
 #' Google Trends data.
 #'