+++ NEWS.md
@@ -1,6 +1,6 @@
 # globaltrends v.0.0.12
 * Stop direct exports from functions to .GlobalEnv
-* Exports of objects (e.g., SQL connections, keyword tables) are redirected to package namespace.
+* Exports of objects (e.g., SQL connections, keyword tables) are redirected to package environment `gt.env`.
 
 # globaltrends v.0.0.11
 * Add function plot_map()